pyproject.toml
[project]
name = "hackathon-2-phase-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "pydantic-core>=2.27.2",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
